Protected signin completed and demonstrated with the FAQ page Moved api endpoints into routes/apiRoutes.js to reduce sprawl in weconnect-server.js
