+++ package.json
@@ -34,7 +34,7 @@
     "@passport-js/passport-twitter": "^1.0.8",
     "@popperjs/core": "^2.11.8",
     "@prisma/client": "^6.0.1",
-    "axios": "^1.7.4",
+    "axios": "^1.7.9",
     "body-parser": "^1.20.3",
     "bootstrap": "^5.3.2",
     "chart.js": "^4.4.0",
